initial commit with finalized code
I wrote the code for this without commits, my apologies if that is against the brief for the project. Adjustment to HTML specifically to change input type "submit" to "button", as well as changing some class or ID names to suit my own preferences. JS code written for the project all before initial commit.